Updating iOS to latest version of my game framework
The latest version of the framework is used in Battle Bombs and
Insectoid Defense. One of the main things I enjoyed about it was that
it allowed me to remove the DirectXTK framework and write graphics code
in a way that was platform agnostic for Android, iOS, and Windows Phone.